[IMP] stock(_picking_batch): add Zero Quantity Count
Whenever a location is theoretically empty, the user can now either
confirm the location is actually empty or adjust the inventory. This
supports keeping a more consistent inventory during the day-to-day work
by making small inventory adjustments from time to time instead of
doing big inventory adjustments more episodically.

This commit implements this feature for the backend (stock only app) by
adding a setting in the inventory settings, and adding a Zero Quantity
Count wizard that will called during the picking validation process.

When the validated pickings empty a location(s), this wizard is raised
and asks for a confirmation of the location(s) empty status, but also
offer the possibility to make an inventory adjustment for those
locations. This is why pickings have now a field `inventory_ids`, those
inventory will be validated right after the picking is validated.

Extra notes:
- The `theoretrical_qty` and the `difference_qty` column of the
  inventory lines are hidden if the inventory adjustment is made
  from a Zero Quantity Count since they will be inaccurate in this case.
- The env.context is cleaned (of default values) due to potential bugs
  that can pop up from leftover default values (e.g. the "Delivery"
  button from a sales order).
- There are some limitations to this implementation such as the ability
  for a user to save => close (without validation) of a zqc triggered
  inventory adjustment (this can lead to inconsistent data).
- If a move occurs after a zqc triggered inventory adjustment is
  created, but before it is validated, this may also lead to
  inconsistent data.
- The same inventory adjustment isn't re-used if one is already saved
  and closed in order to support barcode app functionality.
- Batch pickings are already ideally set up for this to automatically
  work without any changes. Only tests added related to this module.

ENT PR: odoo/enterprise#8932
TaskID: 1938108