[IMP] survey: add background option per section
This commit adds the background per section feature.
If the section to display has a background image configured, the background
will be refreshed using that section background image.
If the section has no background, the survey background is used.
The background is refreshed only if the previous and the next sections are
different. The background is refreshed at the same time that the next question
is loaded.

The next section to display depends on free text (section with description)
configuration and on conditional questions.
The next section to display can be the next question's section or directly
the next question itself (if the question is a section).

Task ID: 2225393