[FIX] survey: don't compute last question based on conditional questions in session manager
As we cannot know what will be the last question before all user inputs are
submitted (because of conditional questions, the last question is still
considered as the last potential question to display.

If, at go next, there is no more question to display, because last questions
are inactive, the final leaderboard will anyway be displayed (already handled
in js controller).

Task ID: 2225393