[IMP] delivery: do not put defaults on non stored fields...
The default value is computed for each record creation even if the value 
isn't stored...
This means in this case, two ref (for the two defaults) are triggered 
for each record created (1 ref = 1 query to verify record existence), 
for nothing.

While on it, batch the computation to avoid useless calls to ref in a 
loop if the field is shown in multi records views.