[FIX] mass_mailing: fix the test send mail button
Before when an user clicked on the button "Test", jinja syntax
was not taken into account for the test mail sent. So if an error
was present, clicking on the button didn't raised an error before
the cron was executed.
Now, if there is atleast one record in the mailing model then, the
template is correctly rendered and throws an error if there is a
syntax error.

task-2312442