[REF] point_of_sale,pos_restaurant: use export_for_ui in generating order json
In this commit, we replace the convoluted way of generating order json that is
used in getting synced orders from the server. We replace it by simply extending
_export_for_ui of core point_of_sale.

In fact, adding fields in the order's json from other module is very difficult
esp. if the module where the field needs to be added is not dependent on
pos_restaurant. Lifting this functionality to the core point_of_sale makes it
easier to add fields in the json by just inheriting _export_for_ui method.