[FIX] base,calendar: fix calendar reminder
The purpose of this commit is to have reliable calendar event reminder.

Before, the reminders was not send directly and was instead put in
mailing queue, which was leading to some delay.
Additionaly, when the cron was executed manually, we never receive the
lastcall date which prevented the system to send reminder when he should.

task-2331999