[IMP] hr_recruitment: Improve the template by Refusing Reason
Purpose of this task is to improve the refuse reason action for sent
the template automatic.

Currently, when using the refuse action,there is no mail template sent to the
applicant.

In this commit we improve the Refuse Reasons view by adding the email template
field and send the template to the related partner by refusing reason action.
and create the new template for not interested applicant.

TaskID - 2313402