[FIX] project: Fix bug
Restore Project Kanban view for mobile.
Fix access denied for Project overview.
Added a name for the default stage of the tasks in the Internal project
Removal of redundancy for the creation of the Internal project and its tasks
Fix task_reccurrence for the type 'after'. (too much task created or bad task created)

Task-2326258