[FIX] web: readonly attribute does not work on boolean_favorite widget
Applying readonly attribute on 'boolean_favorite' field should not display
'Add to Favorites' and 'Remove from Favorites' tip and click on that should
not change the value as well.

With this commit, we are not allowing to click and display tooltip if the field
having 'boolean_favorite' field widget and readonly attribute.

Task : 2339987

Co-authored-by: Mohammed Shekha <msh@odoo.com>