[FIX] website_event_track_online: fix sponsor.image_128 attachment
image_128 field is defined in website_event_track module on sponsor model
as store=True. That means that when this binary field is set, an attachment is
created in ir.attachment model -> attachment is set to True.

When modifying this field in website_event_track_online module and setting
this field into store=True, the attachments are still existing in the
ir.attachment for those sponsor records.

This is not an issue when trying to access the image_128 field via the ORM,
as the compute (or related) is correctly interpreted, but the route
'/web/image' does checks if the field has an attachment, instead of checking
if the field is related to another one. As it is still the case, this route
returns a potentially old attachment.

This commit force the attachment to False when setting the field to
store=False.

This commit also adds a post_init_hook to clean the remaining attachments
linked to the event.sponsor model and 'image_128'.

Task ID: 2341108