[IMP] account: implement carry over functionnality
In a belgian context :
If credit notes amounts is higher than invoices amounts in a grid of the
VAT report, then the amount in the VAT report will be negative. But it
must be forced to 0 in the XML ;
the negative amount that is not included in the XML file must then be
carried over on the next periods. In order to carry over these amounts
to the next period, we will use dedicated analytic accounts.

This process repeats itself from one period to the other endlessly,
until the VAT base amount of these credit notes is fully deducted from
the VAT base amount of the invoices.

The credit note amounts are reported in the effective period in any
case. These amounts are positive, be it the "VAT base amount" or
the "VAT amount".

task id #2271978