[IMP] mail: open partner/employee form on chat header click
PURPOSE:

It often happens that someone will send you a message, but you don't actually
know who is the person within the company. In this context, it would be
practical to have direct access to that person's public user or employee form
in order to get more information.

SPECIFICATIONS:

on hover, display the name of the user is bold + underlined
if the user clicks on the name, redirect to:
 - the employee form if the hr module is installed and if the user is
   linked to an employee
 - else redirect to the partner form

LINKS:

closes https://github.com/odoo/odoo/pull/58754
task 2252924