[FIX] sale: autoselect single radio choice so custom work
When a product is set on a line and we switch, we will use the
attributes values of the previous product.

But if a product value is a radio button with only one option and a
custom value, it will never be selectable if it doesn't match the
previous product attribute.

In this commit, we do not use the current variant values when we change
product.

note: this commit also increases the limit of (custom values) inside the
product configurator wizard, for an arbitrary 10000 limit. Without this
change, each opening of the wizard would only keep 40 customs values and
discard the other ones.

opw-2333091
closes #59016