[IMP] stock: add cyclic inventory for locations
Add a new feature to allow automated (as part of scheduler) generation
of inventory adjustments based on user set value for how often they should
be done for 'internal' and 'transit' locations. When set to a non-zero
value, the next date an inventory should be automatically generated for
a location is based on whether or not a previous inventory for that
location was completed. If an inventory is overdue based on the previous
inventory date, one will be scheduled for tomorrow. If no previous
inventory date exists, schedule the next one based on today's date.
This automation will also delete existing draft inventories that include
this location.

Note that location hierarchies are taken into account (i.e. children
locations will not have an inventory created if their parent already has
an inventory created at the same time.)

Task: 2336455