[IMP] stock: add auto inventory for negative quantities
Add a new feature to automatically create a product/location specific
inventory when its quantity falls below 0. When the quantity reaches
>= 0, auto-delete any of these auto-generated inventories when they are
still draft. The goal of this is to help track inventory discrepencies
and prevent appearance of being out of stock when there is still stock
left.

Scheduler has been updated to check for these auto-generated inventories
and refresh their inventory date if they are still draft and their
quantity is still 0. Scheduler will not check for these auto-generated
inventories if their quantity goes above 0 since this is expected to
already be done when the quantity is updated.

Task: 2336455