[IMP] survey: improve layout of live session in safari(7.1+)
Prior this commit, min-height given on the `o_survey_wrap` class was not
effective on safari browser. This led to different layout of quiz on
safari than other browsers where the content was not centerd vertically.

This commit fixes the issue by providing safari specific css that fixes
the layout for safari browser.

TaskID - 2345907