[FIX] mass_mailing: don't add `opt_out` by default on domain
In mass mailing (and so sms marketing) form view, when you set
`Mailing Contact` as a model, by default the `opt_out` field
is added to domain. However, it's a computed field, and is to
be used only in the context of mailing lists. Otherwise, it
raises the warning (wherever this view is used).

This commit removes `opt_out` field being set by default and
also removes the UserError raised by the search function, thus
avoiding unwanted warnings.

taskId - 2300427