[FIX] web_editor: fix footer/header item move
Using the handle to move elements from footer/header everywhere
in the page will make the footer element disapear with endless
content loading animation.

The end of "drag and drop" triggers the "_onDragAndDropStop"
handler which executes a serie of methods :

_destroyEditors > cleanForSave > toggleTargetVisibility > onTargetShow

The click event triggered on the visibility option in "onTargetShow"
method is only used to toggle footer visibility in the page when we
make it appear for edition.

To prevent this behavior after a "drag and drop", we add a boolean :
"isDragAndDropAction" to control the "onTargetShow" callback.

task-2312878