[IMP] web_tour: tour tip highlighted with a border
Border is fading between Odoo brand light and dark gray: just making it
white did not help noticing it when it was displayed on light
backgrounds.
Tried to make the border size change but had to roll back because a
compensating translate was not exactly matching and the rendered tip was
vibrating.
Translate was adapted to compensate for the change in border size - in
order to keep the position aligned with the intended target.

task-2357170