[Fix] google_calendar: fix owner attendee being deleted
Currently when creating a calendar.event with google sync switched
on and without adding another attendee, the event is immediatly
removed from the calendar view. This is because there is
no attendee_ids in the response from google when there is
only the owner defined on an event.

Fixing this implies adding the owner email to the attendees of
the current event. This is the same thing that is done
in microsoft_calendar, which is available from 14.0 onwards.