[IMP] survey: allow host to go to previous question in session
PURPOSE
=======
For now, even if the setting "Back Button" is activated,
the host, can't go back to the previous screen on a live session.

Why would he want to go back? Misclick, want to display again the result, other

SPECIFICATION
=============
On a live session, let the host go back to the previous screen.
Do not "restart" the questions but display the graph with the results.

Task ID: 2351759